requirements.txt
algoliasearch-django>=2.0,<3.0
django==4.1.4
djangorestframework==3.12.4
psycopg2-binary==2.9.5
requests==2.28.2
pillow==9.4.0
django-guardian==2.4.0
rest-framework-generic-relations==2.1.0
djangorestframework-simplejwt==5.2.2
django-rest-knox==4.2.0
django-review==1.10.0
#django_q==1.3.9